UPSTREAM: dt-bindings: exynos_dsim: update of graph bindings
Of-graph bindings should describe ports present in the device, not the
devices it can be connected to. The patch replaces verbose description
with shorter but more precise one. While at it clock related properties
are moved to the main node as it is their actual location.

Signed-off-by: Andrzej Hajda <a.hajda@samsung.com>
Signed-off-by: Inki Dae <inki.dae@samsung.com>

(cherry picked from commit 91e28030fd5deb56b5f14b870bc2e2c856435cca)

BUG=None
TEST=Tested on cheza, build tested everything else

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Change-Id: Ief67a715be3541bfe710202c898600d62095caba
Reviewed-on: https://chromium-review.googlesource.com/1342158