UPSTREAM: OPP: Don't take OPP table's kref for static OPPs
The reference count is only required to be incremented for every call
that may lead to adding the OPP table. For static OPPs the same should
be done from the parent routine which adds all static OPPs together and
so only one refcount for all static OPPs.

Update code to reflect that.

The refcount is incremented every time a dynamic OPP is created (as that
can lead to creating the OPP table) and the same is dropped when the OPP
is removed.

Tested-by: Niklas Cassel <niklas.cassel@linaro.org>
Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>

BUG=b:80592702
TEST=Build and boot

Change-Id: Ifec16d09e8605ed9d79da60997e27f68290ffcd6
Signed-off-by: Rajendra Nayak <rnayak@codeaurora.org>
(cherry picked from commit 0ad8c623907c27f4b8572d36c4ba73ea103e3108)
Reviewed-on: https://chromium-review.googlesource.com/1339303
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>