UPSTREAM: mfd: cros_ec: Check for mkbp events on resume only if supported.
Currently on every resume we check for mkbp events and notify the
clients. This helps in identifying the wakeup sources. But on devices
that do not support mkbp protocol, we might end up querying key state of
the keyboard in a loop which blocks the resume. Instead check for events
only if mkbp is supported.

Signed-off-by: RaviChandra Sadineni <ravisadineni@chromium.org>
Reported-by: Marek Szyprowski <m.szyprowski@samsung.com>
Tested-by: Marek Szyprowski <m.szyprowski@samsung.com>
Signed-off-by: Lee Jones <lee.jones@linaro.org>
(cherry picked from commit 61cc15dac01ae84281222452e338ca060179d8b1)

BUG=chromium:941638
TEST=Suspend/resume doesn't hang anymore

Change-Id: I3914077c8feae2025ed98443b8ea2f958160a3e6
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1549746
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Ravi Chandra Sadineni <ravisadineni@chromium.org>