FROMGIT: dt-bindings: firmware: scm: Refactor compatibles and clocks
When the binding was written all "future" platforms required three
clocks, so the default compatible (qcom,scm) was defined to require
this. But as history shows all "future" platforms actually lack required
clocks. Given how the binding is written these compatibles have to be
added as an exception to the default.

Refactor the description of compatible to define that a platform
compatible should be given, followed by the fallback of qcom,scm. Also
refactor the description of the clocks in a way that this does not need
to be updated as new platform specific compatibles are added.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Reviewed-by: Stephen Boyd <sboyd@kernel.org>
Signed-off-by: Andy Gross <andy.gross@linaro.org>
(cherry picked from commit 8a07855e66e6b8ddf452d81c6aac0b1ff3665e86
 https://git.kernel.org/pub/scm/linux/kernel/git/agross/linux.git for-next)

BUG=b:118134560
TEST=Boot cheza

Change-Id: Iecbd56abb841f71c6e97b8ab2dab6867fbf5f039
Signed-off-by: Evan Green <evgreen@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1305182
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>